refactor: clean up code organization
